+++ src/components/Grid/index.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
-import GridCol from './components/GridCol';
+import GridCol from './GridCol';
 
 const Grid = styled('div').attrs(() => ({
   className: 'f-Grid',
