+++ src/components/UiText/index.stories.js
@@ -15,7 +15,7 @@ stories.add('All states', () => (
     <Heading>UiText</Heading>
     <Heading level={2}>Variants</Heading>
     {variants.map(variant => (
-      <UiText key={variant} variant={variant} as={variant}>
+      <UiText key={variant} variant={variant}>
         {variant}
       </UiText>
     ))}
